+++ packages/pyxle-db/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2026-09-05
+
 ### Changed
 
 - `__version__` and the plugin's `version` are read from installed distribution